+++ mlp/lossf.h
@@ -6,7 +6,7 @@
  * @param  label `[IN]`输出层标签
  * @return 损失值
  */
-double mse_loss(Vector *out, Vector *label);
+float mse_loss(Vector *out, Vector *label);
 
 /**
  * @brief  计算输出层梯度（平方差损失函数）
@@ -22,7 +22,7 @@ Vector *d_mse_loss(Vector *out, Vector *label);
  * @param  label `[IN]`输出层标签
  * @return 损失值
  */
-double ce_loss(Vector *out, Vector *label);
+float ce_loss(Vector *out, Vector *label);
 
 /**
  * @brief  计算输出层梯度（交叉熵损失函数）
@@ -38,7 +38,7 @@ Vector *d_ce_loss(Vector *out, Vector *label);
  * @param  label `[IN]`输出层标签
  * @return 损失值
  */
-double softmax_ce_loss(Vector *out, Vector *label);
+float softmax_ce_loss(Vector *out, Vector *label);
 
 /**
  * @brief  计算输出层梯度（归一化指数函数 + 交叉熵损失函数）
